Add support for sitemap index using fork of SitemapGen4j
I created this fork to customize the generation of the sitemap index
by SitemapGen4j. There were several issues with the previous behavior:

  * the limit of the maximum number of URLs per file had to be detected by
    client code to call the method writeSitemapsWithIndex() of
    the WebSitemapGenerator object [1]

  * the name of the generated sitemap was fixed to "sitemap_index.xml",
    which differs from the name "sitemap.xml" of a single sitemap.
    This means that the robots.txt file had to modified to declare
    either "sitemap.xml" before the maximum number of URLs is reached
    and "sitemap_index.xml" after

  * the index could not be generated systematically either, because the
    generation failed below the threshold of the maximum number of URLs

In the fork implementation, the sitemap index is now generated automatically
when the sitemap is cut into several files, with the name "sitemap.xml"
to keep the same entry point declared in robots.txt. The generated sitemaps
are still called "sitemap1.xml", "sitemap2.xml", ... like in the original
project. I also updated the XML schemas used for validation, which prevents
unexpected date format errors when generating sitemap indexes on Linux.

[1] SitemapGen4j - project home
http://code.google.com/p/sitemapgen4j/

[2] Issue 8:  wsg.writeSitemapsWithIndex()
http://code.google.com/p/sitemapgen4j/issues/detail?id=8