Remove generated sitemaps from test-resources/ in clean task
After running one of the tasks use.* manually (these tasks are also
called in unit tests, with proper cleanup) the generated sitemaps
cause the first unit test that runs afterwards to fail.

Running 'ant clean' before 'ant test' solves this issue.

I now consider adding "clean" to the dependencies of the "test" task,
or running the tearDown() method before the first run of the unit tests
to complete this fix.